requirements.txt
selenium~=4.31.0
allure-python-commons~=2.14.0
behave~=1.2.6